chore: share points overview components between desktop / mobile
