+++ apps/marginfi-v2-ui/src/components/common/Points/index.tsx
@@ -1,0 +1,2 @@
+export * from "./PointsConnectWallet";
+export * from "./PointsOverview";
