+++ apps/marginfi-v2-ui/src/components/desktop/Points/index.ts
@@ -1,3 +1,1 @@
-export * from "./PointsOverview";
 export * from "./PointsTable";
-export * from "./PointsConnectWallet";
