--- apps/marginfi-v2-ui/src/components/mobile/Points/index.ts
@@ -1,1 +1,0 @@
-export * from "./MobilePointsOverview";
