--- apps/marginfi-v2-ui/src/components/mobile/Points/style.module.css
@@ -1,9 +1,0 @@
-.hide-scrollbar {
-  -ms-overflow-style: none; /* for Internet Explorer, Edge */
-  scrollbar-width: none; /* for Firefox */
-  overflow-x: scroll;
-}
-
-.hide-scrollbar::-webkit-scrollbar {
-  display: none; /* for Chrome, Safari, and Opera */
-}
